+++ src/style-spec/CHANGELOG.md
@@ -1,3 +1,7 @@
+## main
+
+* Implement `replace-all` expression. ([#2059](https://github.com/maplibre/maplibre-gl-js/pull/2059))
+
 ## 17.0.2
 
 * Fix errors when running style-spec bin scripts and added missing help. Removed unnecessary script 'gl-style-composite'. ([#1971](https://github.com/maplibre/maplibre-gl-js/pull/1971))